# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a project made by CCNU 书斋",
        "title": "2021-miniproject-Library",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/Library/:books_id": {
            "get": {
                "description": "\"从书城进入书籍\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "\"书籍信息\"",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BooksInfo"
                        }
                    },
                    "401": {
                        "description": "查找失败"
                    }
                }
            }
        },
        "/Library/:books_id/digest": {
            "get": {
                "description": "\"从书城进入书本获取的关于这个书的书摘信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "\"书本的书摘信息\"",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DigestInfo"
                        }
                    },
                    "404": {
                        "description": "获取失败"
                    }
                }
            }
        },
        "/Library/addbook/:books_id": {
            "post": {
                "description": "\"通过书本上面的添加按钮来实现添加\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "\"添加书本到我的书架\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "toekn",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功"
                    },
                    "400": {
                        "description": "添加失败"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            }
        },
        "/Library/seacher": {
            "post": {
                "description": "\"在书城中搜索书本\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "\"搜索书本\"",
                "responses": {
                    "200": {
                        "description": "搜索成功",
                        "schema": {
                            "$ref": "#/definitions/model.BooksInfo"
                        }
                    },
                    "401": {
                        "description": "请重试"
                    },
                    "404": {
                        "description": "搜索不到"
                    }
                }
            }
        },
        "/digest": {
            "post": {
                "description": "将创建的新书摘加入数据库中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digest"
                ],
                "summary": "添加书摘",
                "responses": {
                    "200": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "编辑错误"
                    }
                }
            }
        },
        "/digest/mysummary/:user_id": {
            "get": {
                "description": "\"具体信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digest"
                ],
                "summary": "\"书摘主界面\"",
                "responses": {
                    "200": {
                        "description": "获取成功"
                    }
                }
            },
            "put": {
                "description": "\"通过按钮来确定是否公开这个书摘\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digest"
                ],
                "summary": "\"是否公开\"",
                "responses": {
                    "200": {
                        "description": "修改成功"
                    },
                    "401": {
                        "description": "修改失败"
                    }
                }
            },
            "post": {
                "description": "\"搜索书摘的信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digest"
                ],
                "summary": "\"搜索\"",
                "responses": {
                    "200": {
                        "description": "获取成功"
                    },
                    "400": {
                        "description": "搜索错误"
                    }
                }
            }
        },
        "/digest/mysummary/:user_id/clasees_edit": {
            "delete": {
                "description": "\"删除我的书摘分类里的类别\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digest"
                ],
                "summary": "删除书摘的分类",
                "responses": {
                    "200": {
                        "description": "成功删除"
                    }
                }
            }
        },
        "/digest/mysummary/:user_id/classes": {
            "get": {
                "description": "\"书摘的各种类别\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digest"
                ],
                "summary": "\"书摘分类\"",
                "responses": {
                    "200": {
                        "description": "获取成功"
                    }
                }
            }
        },
        "/digest/mysummary/:user_id/classes_edit": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digest"
                ],
                "summary": "\"获取我的书摘分类\"",
                "responses": {
                    "200": {
                        "description": "获取成功"
                    }
                }
            }
        },
        "/digest/person/:summary_id": {
            "put": {
                "description": "\"通过编辑修改我发布的书摘\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digest"
                ],
                "summary": "\"编辑书摘\"",
                "responses": {
                    "200": {
                        "description": "编辑成功"
                    },
                    "400": {
                        "description": "编辑错误"
                    },
                    "404": {
                        "description": "编辑失败，数据为空"
                    }
                }
            }
        },
        "/homepage/:user_id": {
            "get": {
                "description": "\"通过我的id进入我的主页\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "my"
                ],
                "summary": "\"我的主页\"",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"success\"}",
                        "schema": {
                            "$ref": "#/definitions/model.Userinfo"
                        }
                    },
                    "404": {
                        "description": "找不到该用户的信息"
                    }
                }
            }
        },
        "/homepage/:user_id/info": {
            "put": {
                "description": "\"修改用户的基本信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "my"
                ],
                "summary": "\"修改用户的信息\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "toekn",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功"
                    },
                    "400": {
                        "description": "修改失败"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            }
        },
        "/homepage/:user_id/mydigest": {
            "get": {
                "description": "\"显示我的发布中的书摘的有关信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "my"
                ],
                "summary": "\"我的发布\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "toekn",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"success\"}",
                        "schema": {
                            "$ref": "#/definitions/model.DigestInfo"
                        }
                    },
                    "404": {
                        "description": "获取失败"
                    }
                }
            }
        },
        "/homepage/:user_id/mydigest/digest_id": {
            "get": {
                "description": "\"获取我的书=书摘的具体信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "my"
                ],
                "summary": "\"查看我的发布中书摘的具体信息\"",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DigestInfo"
                        }
                    },
                    "401": {
                        "description": "获取失败"
                    }
                }
            },
            "put": {
                "description": "\"删除我的发布中的书摘\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "my"
                ],
                "summary": "\"删除我的发布\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "toekn",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "401": {
                        "description": "删除失败"
                    },
                    "404": {
                        "description": "认证失败"
                    }
                }
            }
        },
        "/homepage/:user_id/shelf": {
            "get": {
                "description": "\"通过用户的id获得该用户的书架上书本的信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "my"
                ],
                "summary": "\"我的书架\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "toekn",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"success\"}",
                        "schema": {
                            "$ref": "#/definitions/model.BooksInfo"
                        }
                    },
                    "404": {
                        "description": "获取失败"
                    }
                }
            }
        },
        "/homepage/:user_id/shelf/book_id": {
            "get": {
                "description": "\"查看我的书本的具体信息\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "my"
                ],
                "summary": "\"我收藏的书\"",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BooksInfo"
                        }
                    },
                    "401": {
                        "description": "查找失败"
                    }
                }
            }
        },
        "/homepage/:user_id/shelf/books_id": {
            "put": {
                "description": "\"删除数据库中这本书与该用户的关系\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "my"
                ],
                "summary": "\"删除我的书架上的书\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "toekn",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "401": {
                        "description": "删除失败"
                    }
                }
            }
        },
        "/library": {
            "get": {
                "description": "\"显示书城主界面\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "\"书城\"",
                "responses": {
                    "200": {
                        "description": "获取成功"
                    }
                }
            }
        },
        "/library/class/:class_id": {
            "get": {
                "description": "\"显示该类别下的书籍\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "\"分类的书籍\"",
                "responses": {
                    "200": {
                        "description": "获取成功"
                    }
                }
            }
        },
        "/library/classes": {
            "get": {
                "description": "\"书本的各种类别\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "\"书籍分类\"",
                "responses": {
                    "200": {
                        "description": "获取成功"
                    }
                }
            }
        },
        "/library/ranking": {
            "get": {
                "description": "\"通过点击量对书本的热度进行排序\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "\"书城排行榜\"",
                "responses": {
                    "200": {
                        "description": "获取成功"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "\"验证用户信息实现登入\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "\"登入\"",
                "responses": {
                    "200": {
                        "description": "登入成功",
                        "schema": {
                            "$ref": "#/definitions/model.Token"
                        }
                    },
                    "400": {
                        "description": "输入格式错误"
                    },
                    "401": {
                        "description": "密码错误"
                    },
                    "404": {
                        "description": "用户不存在"
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "\"注册一个新用户\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "\"注册\"",
                "responses": {
                    "200": {
                        "description": "注册成功"
                    },
                    "400": {
                        "description": "输入有误，格式错误"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.BooksInfo": {
            "type": "object",
            "properties": {
                "book_author": {
                    "type": "string"
                },
                "book_id": {
                    "type": "string"
                },
                "book_information": {
                    "type": "string"
                },
                "book_name": {
                    "type": "string"
                },
                "book_picture": {
                    "type": "string"
                },
                "class_id": {
                    "type": "string"
                },
                "click_sum": {
                    "type": "integer"
                }
            }
        },
        "model.DigestInfo": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "string"
                },
                "class_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "summary_information": {
                    "description": "摘录的原文",
                    "type": "string"
                },
                "thought": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "model.Userinfo": {
            "type": "object",
            "properties": {
                "motto": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_password": {
                    "type": "string"
                },
                "user_picture": {
                    "type": "string"
                }
            }
        }
    }
}